pyproject.toml
[tool.poetry]
name = "asllvd-skeleton-creator"
version = "0.1.0"
description = ""
authors = ["Cleison Amorim <cca5@cin.ufpe.br>"]

[tool.poetry.dependencies]
python = "^3.7"
commons-python = { git = "https://github.com/amorim-cleison/commons-python.git", branch = "develop" }
numpy = "^1.19.2"
pandas = "^1.1.3"
ffmpy = "^0.2.3"
PyYAML = "^5.3.1"
torchlight = "^0.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
yapf = "^0.30.0"
pytest = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
